Fix for ugly hardcoded French 404 page
In case of a missing controller or plugin, we would have received an ugly French error message that was not wrapped in a layout and did not offer any chance of styling. Now uses the configured renderer to display the error page, allowing full styling of the error and as a bonus, it now checks the 'skipLog' parameter of the Exception configuration to find out which exceptions should be excluded from Sentry.